feat: 없는 예약 또는 지나간 예약 삭제 요청시 예외 발생 기능 구현

+++ src/main/java/roomescape/service/ReservationService.java
@@ -105,6 +105,14 @@ public class ReservationService {
 
     @Transactional
     public void delete(Long id) {
+        Optional<Reservation> reservation = reservationRepository.readById(id);
+        if (reservation.isEmpty()) {
+            throw new CustomException(ErrorCode.NOT_FOUND_RESERVATION);
+        }
+        if (reservation.get().isPast(LocalDateTime.now())) {
+            throw new CustomException(ErrorCode.PAST_TIME_RESERVATION_DELETE);
+        }
+
         reservationRepository.delete(id);
     }
 }
